move event handlers to directory
