+++ backend/app/controllers/event/deleteEvent.js
