+++ backend/routes/api.js
@@ -4,9 +4,9 @@ const router = express.Router();
 
 const listCalendars = require( '../app/controllers/calendar/listCalendars' );
 const getFreeCalendars = require( '../app/controllers/calendar/getFreeCalendars' );
-const fetchCalendar = require( '../app/controllers/calendar/fetchCalendar' );
-const addEvent = require( '../app/controllers/calendar/addEvent' );
-const deleteEvent = require( '../app/controllers/calendar/deleteEvent' );
+const fetchCalendar = require( '../app/controllers/event/fetchCalendar' );
+const addEvent = require( '../app/controllers/event/addEvent' );
+const deleteEvent = require( '../app/controllers/event/deleteEvent' );
 
 router.get( '/calendars', listCalendars );
 router.post( '/calendars/free', getFreeCalendars );
